* solaris-fixes:
  build: Include <sys/loadavg.h> if it exists.
  lib-tests: test-io{cache|broker} need -lsocket to link.
  lib-tests: test-runcmd assumes GNU echo.
  lib-tests: Signal handlers don't return int on most platforms, and using a Linux-specific internal typedef to cast away the incompatible pointer warning when registering with signal() was the wrong way to resolve this.
  Fix some type/format mismatch warnings for pid_t.
  Fix build on Solaris.

+++ lib/test-iocache.c
@@ -50,8 +50,8 @@ static int test_delimiter(const char *delim, unsigned int delim_len)
 				   delim_len, i, len, sc[i].len);
 			printf("sc[i].str: %s\n", sc[i].str);
 			printf("ptr      : %s\n", ptr);
-			printf("strlen(sc[i].str): %lu\n", strlen(sc[i].str));
-			printf("strlen(ptr)      : %lu\n", strlen(ptr));
+			printf("strlen(sc[i].str): %lu\n", (unsigned long)strlen(sc[i].str));
+			printf("strlen(ptr)      : %lu\n", (unsigned long)strlen(ptr));
 			exit(1);
 		}
 	}
